float

+++ GC_UE4CPP/Source/GC_UE4CPP/Chest.h
@@ -32,7 +32,7 @@ public:
 	UFUNCTION()
 		void OnOverlapEnd(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex);
 
-	int compteur = 0;
+	float compteur = 0;
 
 
 protected:
